requirements.txt
Flask~=2.0.2
itsdangerous~=2.0.1
alembic==1.7.4
aniso8601==9.0.1
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cryptography
dnspython==2.1.0
email-validator==1.1.3
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-HTTPAuth==4.5.0
Flask-Mail==0.9.1
flask-mongoengine==1.0.0
Flask-RESTful==0.3.9
Flask-WTF==0.15.1
greenlet
gunicorn==19.7.0
httplib2==0.20.2
idna==3.3
Jinja2
Mako==1.1.5
MarkupSafe
mongoengine==0.23.1
olefile
passlib==1.7.4
Pillow==8.4.0
pycparser
pymongo==3.12.0
pyparsing==3.0.4
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
requests==2.26.0
six==1.16.0
urllib3==1.26.7
Werkzeug
wincertstore==0.2
WTForms==2.3.3

